bits 12 & 13 of GDKdebug (4096 & 8192) were already used by gdk for BBPPerformance debugging & Delta debugging, respectively, hence, we assign bits 14 & 15 (16384 & 32768) to Module loading & Yacc specific error messages, instead.
bit 16 (65536) seems to be used by tcpip, and
bit 17 (131072) seems to be used by monet_multiplex for some setaggr debugging.

ddbench used to use bits 14, 15, & 16 (16384, 32768, & 65536);
changed these to bits 18, 19, & 20 (262144, 524288, & 1048576)
to re-establish unambiguity.

fixed semantic swap that creeped in with Martin's changes:
it was "(module|yacc)*silent*", so it should be "(!(GDKdebug&<resp.value>))", now,
not "(GDKdebug&<resp.value>)"