The void cleanup (These changes break things in projects)!
These changes should clearup any old inconsistencies in oids.

On a logical level 2 types exist void and oid, where
	void is the real void, ie no data!.

On the physical level the void atom is misused for compressed storage of
dense oids (can be seen because the seqbase != oid_nil).

Incase of updates these bats are now transformed into oid bats.

Beware that:

	Inserting in a void(seqbase != oid_nil) used to mean auto increment.
	This is nolonger there ie fix your mil scripts and c-code!

Also in this version a call to BBPscandisk is made when starting, needed
to cleanup no longer needed bats on disk after quit or crash.
Yes also needed for quit as this currently does no cleanup.

(small) Tests for new void semantics and for the slice mil operator are added.