propagated Menzo's fix from main branch:
PROBLEM:
When processing Makefile.ag in the topdir CWD was set to './' so
configure.in would contain a request for expansion of ./Makefile.in.
configure would generate config.status containing checks for the
target ./Makefile, however, in an incremental Make 'config.status --recheck'
is followed by 'config.status Makefile', but ./Makefile != Makefile leading to
an "invalid argument" error.

SOLUTION:
When cwd == topdir CWD is set to ''