autogen change  libraries can be installed in a given directory example
lib_x {
	DIR = pkglibdir
}
(use of reseting LIBDIR is now depricated)

This is needed to allow sub projects to install also int lib/MonetDB which
is again needed both for easy user installs and for rpm packaging.

added read and write functions to the  atom api.
	.READ = x_read
	.WRITe = x_write

in @m
in @c you need
	x *x_read(x *a, stream *s){
		//read from stream s into a
		// if varsized allocate a new atom a
		// do not free the old *a.
		return a;
	}
	void x_write(x *a, stream *s){
		// write a on to the stream s
	}

as streams can both be ascii and binary stream there should be
an if (s->type == ST_ASCII) but this is currently not implemented
for the basic atoms (such as int/str).