Changed order of checking for iconv; Now, first check for a library iconv containing   either function iconv   or function libiconv If found, add -liconv to Makefile.
If not found, check if iconv is still known (e.g. by libc),
in which case have_iconv is yes but no library flags are set.

As a result, a standalone iconv library that has been installed
takes precendence over the "standard" iconv functions in [g]libc.

(Note: this was a setup issue needed on my RH6.2 (still alive :) home system).