Merged recent changes on main branch into CWI_DDI_merge branch.
Many tests work correctly.  Some give differing but expected output,
some give output that needs looking in to, and some are still wrong.
But it mostly works.

The revision up to which was merged into this branch is tagged
merged_to_CWI_DDI_branch; the revision in the branch which contains
the merge (i.e. after this commit) will be tagged
merged_from_head_branch.