the BBPtrim, but also other BAT-loading critical phases actually want to wait for all pending unloads to be finished. i.e. if memory is short, an unload that will free a lot of memory could already be underway, but just take a lot of time to complete due to having to save a large BAT.
thus, we want  to be able to wait on all pending unloads. This is done
using two new locks and a small adaptation of the code that switches
of the BBPUNLADING status bit (it now also decrements qthe unload counter
protected by the GDKunloadLock, and releases the GDKunloadBarrier if
the last unload finishes).

There is also a small change here in BBPreclaim. It reterns a success
value and is also more precise in when an error must be given and when not.
BBPreclaim performs a BBPunfix an unloads a BAT (if possible).