three big chanmges in the module administration (drop-related)
safe module drops
=================

in order to make module drops save, we have eliminated the VARpurge method
of (unsafely) overwriting variables, and non-race-dondition-free of unloading
BATs with soo-to-be-obsolote types from the BBP.

first of all, drops can now fail, if a user holds a reference to an
atom-to-be-dropped. so nothing is overwritten.

this is accomplished by a new function unloadmod_check_use() in monet_tbl.mx,
which in turn depends on VARfindinstance.

bats pinned by prologue/unpinnend by epilogue
=============================================

some modules caeate data structures with the C-prologgue and deallocate them
on physical drop with the C-epilogue. Often these are BATs. Problems would
arise if these BATs contain types introduced by the module:

how can the drop know that a BAT with a reference still on it at drop time
(but before the epilogue has been executed) that it is a dangling value
(for which the drop must be blocked) or if it will be deallocated by the epilogue?

the .PROLOGUE routine now returns a bat*, that is a zero-terminated sequence of
BAT-ids. 99% of all modules with a prologue can just return NULL. Only salgebra
at this moment, has to return something here.

so, this led to (small) changes in all modules with a .PROLOGUE

cleaning up MIL load scrips
===========================

Niels/Strefan removed .LOAD/.DROP and replaced it with a MIL-only module structure
that contained the .LOAD script. As there are no .DROP scripts anymore, variables
and procedures introduced by modules are kept dangling.

This is now done by registering which module defines which proc during parsing
(thanks to ll_parent_module() and the stack structure in introduced  by Niels
we now know the MIL code of which module is parsing), during execution when
new CONST and VARs are introduced (this is done by a before-after check of the list
of variables in the root client context, again using the second stack structure
introduced by Niels in monet_client.mx).

main changes in  TBL_newproc/ TBL_delproc (monet_tbl.mx) and CNTXTregvars (monet_context.mx)

Thus when dropping a MIL module (name does not start with underscore),
we can delete all procs and remove all MIL variables. To reduce race conditions
to the minimum, such variables are actually kept alove, but their value changes
type to void, their name is changed into an impossible name, and they are placed
at the end of the variable list (in order not to hinder searching). Reason is that
MIL loop code contains hard pointers to Variable records, and removing a variable
before the context has ended is potentially dangerous (VARdelvar, monet_context.mx).