Slight change of Stefan's latest change:
The first thing Mserver.h does is to include monet.h;
the first thing monet.h does is to include gdk.h;
the first thing gdk.h does is to include monet_utils.h;
the first thing monet_utils.h does is to include config.h.

This means if a .c file wants to include any of the five .h files
mentioned above, it should be the first one it includes (and it should
not include any of the others), since we want config.h to be
(effectively) the first file that gets included.

Also moved some system includes and defines to monet_utils.h.

All of this on the request from Niels so that he can include
monet_utils.h in sql (which he needs) and get config.h and as a bonus
(for which he doesn't care particularly, but which is needed).