extended test modrefcnt2, and added stable output as I would expect it:
clients do see and can access modules (and their atoms)
that were loaded in the global context on the Mserver konsole
before the client connected.

currently, the client does see that the module is loaded,
but cannot access the respective atom.
this is IMHO inconsistent.


See also SourceForge bug report #815433.