By default, we now include "monetdb_config.h" in the extensions' "*_config.h"; thus, we only need to include "*_config.h" in the sources.
Also added a variable "CONFIG_H" to the various configure.ag that tells us
how the local "*_config.h" is called; this allows us to have identical "rules.mk", again.

These changes break backward-compatibility with the stable Monet_4-3-14c,
hence, they are only applied to the SF-hosted extensions,
but not yet to the CWI-hosted ones.