In the MIL test function, print the Unicode characters using the format u%04x instead of %d.  The whole Unicode world uses hexadecimal representation for Unicode code points, so that makes comparing easier.
Use title case instead of upper case for translating the toUpper
function.  This function translates the whole string to upper case, so
title case is appropriate.  (This only matters for a very limited set
of Unicode code points: dz with caron, lj, nj, and dz.)

For the conversion table, use a single array with two fields per entry
instead of two single-field arrays.