Check, check, check!
Check all reads, writes and flushes in stream.mx, and check the return
values of the stream_* functions.  And do something useful with the
result!

Also, ignore SIGPIPE so that writes to closed sockets and pipes just
return a failure (which is checked!).  Use SIGCHILD (if available) to
indicate that we need to wait for children that have exited (instead
of SIGPIPE which has nothing to do with child processes).