In socket_close, close the *read* socket instead of the *write* socket.  This is because in mapi.mx there is otherwise a potential file descriptor leak: if communication fails at a very early stage, the read stream is closed before the write stream is even created.
Also removed superfluous closes of the socket underlying the socket
streams.  They were a race condition waiting to happen.

Also, what kind of code is this?  This is absolutely bad style and has
to be rooted out!  And I mean it.  (Luckily I had to remove some of
these lines.)
if( c->socket) close(c->socket); c->socket=0;