propagated the remaining changes of today from the development branch to the release branch, as they either appear to be "save"/"harmless" (Sjoerd, Menzo), or are declared as bug fixes (Peter); see messages below.
Now, the release branch and the development branch are in sync, again,
and tagged as "MonetDB_4-3-16_rc05" and "MonetDB_4-3-17_05", respectively.

IMPORTANT NOTE:
===============

As of NOW, check-ins may go to *only one* of the two branches:

- All bug fixes have to go to the release branch, only!
  I will take care of propagating them to the development branch
  regularly (say, once a day).

- All new features have to go to the development branch, only!
  All check-ins to the development branch are considered new features.
  No check-ins to the development branch are ever to be propagated
  to the release branch!


===================================================================
2004/03/12 - windhouw: de-bootstrap
de-bootstrap:
- -e option to echo makes "\c" effective
===================================================================
2004/03/12 - sjoerd: RunMserver.bat.in NT/Makefile
Added a BAT script to run an uninstalled Mserver on Windows.
===================================================================
2004/03/12 - sjoerd: NT/winconfig.h
Don't use long long on VS .NET.  The %lld printf format doesn't work.
===================================================================
2004/03/12 - sjoerd: src/utils/autogen/msc.py
Create script files (especiall MIL scripts) while compiling, not just when checking/installing.
===================================================================

===================================================================
2004/03/12 - boncz: src/gdk/gdk_utils.mx
performance fix (attempt)
- GDKms is now only called on a BBPtrim, which should be almost never
- GDKheapinc/GDKheapdec, GDKmeminc/GDKmemdec have become macros
- GDKmeminc/GDKmemdec still lock, but without if to see if it is set
===================================================================
2004/03/12 - boncz: src/gdk/gdk_utils.mx
performance fix:
- get MT_mallinfo out of GDKmemchk (happened on linux)
===================================================================
2004/03/12 - boncz: src/monet/monet_interpreter.mx
Fix for a bug reported by Roberto todae (crash in a paralell batloop that
is in a while-loop)

what happened here is that on the last iteration, if there are less than N
items, the MIL tree is hacked by the parloop code to hide some children
(for those statements where there is no data).

This hack should be undone afterwards, so the expanded tree that contains a MIL
parblock, stays in sane state, as itis re-used if the batloop itself is in
some other loop.
===================================================================