Three bugs conspired together to make such a simple script as
commit;
commit;

hang on Windows with Visual Studio .NET.

The first bug is that TMcommit called BBPlock, but didn't always call
BBPunlock.  In particular, if the call to BBPsync failed, the unlock
was skipped.  This resulted in a second call to TMcommit hanging.

The second bug is that BBPsync in certain circumstances returned
failure, even though the essential part of the operation, the renaming
of BAKDIR to DELDIR, succeeded.  In particular, if removing DELDIR
failed, or if creating a new BAKDIR failed, or if moving BBP.dir from
BATDIR to BAKDIR failed, BBPsync would return failure.  None of those
later failures is serious.

The third bug, which triggered the above bugs, is that aparently in
Visual Studio .NET it is a failure if rmdir and mkdir are called with
a path that ends in \.