Fix a bug which was introduced way back in '97:
When a random number is generated which had been generated in the
past, just try again instead of seeding the random number generator.
The period of the simple PRNG is much longer than the number of random
numbers that we need, so it is not necessary to seed the generator
when we hit a duplicate.  In fact, seeding it with the current time is
actually a guarantee that we will hit lots of duplicate random numbers
since the time is only accurate to the second, and you can call rand()
quite a few times withing a second.