BUGFIX 920100:
The HEAPload would *always* move a X.priv file to a X. This changes the
semantics the 'new' huge heaps, causing among others an assertion in
make_readonly() to go off.

Ah, the real cause of the bug is also fixed here: a heap leak X is taken
over as the heap to load when a new 'huge' heap X.priv is loaded.