another unreported bug..
A new non-committed persistent bat with 'huge' .priv heaps is saved efficiently
with the msync() call. However, when the commit succeeds, its is not legal
anymore to use the STORE_MMAP mode, as you are never sure which updated pages
got flushed to disk in case of a crash.

This problem is addressed by unloading the bat right inside the committ,
after saving it with msync(). However, it may ha[ppen that the BAT is in use
during the commit and coannot be dropped. In that case we cmust change
the heap modes to STORE_PRIV. From the OS perspective, it is still
mmap(STORE_MMAP) memory on the X.priv file, but the MonetDB commit will
save the BAT in an independent X file. Thus, the penalty is a more expensive
write (iso the msync) and duplicated disk storage.

This is implemented in the dirty_bat() routine of BBPsync.  This must be
done because a 'huge' heap bat may even be totally clean. Just the fact that
the commit happensreuquires to make at least its descriptor dirty (for the
modified heap mode).

The logic to determine the legal mmap modes of a BAT heap (previosuly in
gdk_storage.mx DESCsetmodes/HEAPcheckmod) is now moved to an exported function
BATcheckmodes() in gdk_bat.mx. It is used by both dirty_bat() and
DESCsetmodes().

ah, something else: gdk_bbp.mx heap_move():

do not need to backup STORE_MMAP heap. These normally are not upposed to
be dirty, but actually for BAT_APPEND bats they may be. As all modifications
occur past the stable point, backuops are not necessary.