BUGFIX 924909
this is a race condition. BBPfree saves a BAT, and then removes it
from the BBPcache (BBPclear, requiring a GDKCachelock and/or BBPuncacheit,
requiring a GDKswapLock). After BBPfree, unloadBarrier is released.

(unloadBarrier is taken as long as at least one BAT is being unloaded --
 this can take a long time as the BAT may need to be saved to disk).

TMcommit first gets all the GDKcacheLock, then all GDKswapLocks
and then waits for unloadBarrier

The unloadBarrier exists just because in the commit we want to
make sure that descriptors do not disappear during the commit because
a BAT finishes saving. BBPcyns (ie commit) gets all locks, blocks BBPtrim,
and assumes that the BBP loaded state is stable while it runs.

the deadlock is as follows:
TMcommit:                         BBPtrim(b):
                                    set BBPUNLOADING bit in BBP_status(b)
                                    get unloadBarrier
                                    BBPfree(b)
                                    +BBPsave
 get cachelock,swaplocks            |
 get unloadbarrier ==> BLOCKS!      |
                                    +BBPclear(b)
                                     get cacheLock ==>BLOCKS!

the resolution is to release unloadBarrier from within BBPfree after
the save has been done but before the BBPclear

However, this would mean that BBPclear would remove the bat from BBP_cache
during the commit. That defeats the purpose of unloadbarrier.

Therefore a BBPuncacheit in BBPfree must happen before giving
back the barrier (after giving back the barrier, this BBP_cache() may not
be modified anymore).

But, BBPuncachit took a BBPswapLock! This would have caused another deadlock,
as TMcommit, blocked on unloadbarrier, already has all swaplocks.

The way out is to realize that BBPuncacheit does not need the swapLock at all.
You may read from BBP_cache() without locking, but *only* if you have got
a BBPfix on the BAT. You can never have a BBPfix while a BAT is unloading.
So we know that nobody is reading BBP_cache(bid) anyway during a BBPuncacheit.

summary of modifications:
- give back unloadBarrier from within BBPfree instead of after it
- remove locking from BBPuncacheit