ResultSet seemed not to compile?!?
Made a little experimental change to the JdbcClient which changes the
way it tries to log in. Now looks for a ~/.monetdb file for username,
password, hostname, port and (the not used) database.
Command line options override these preferences from file, and the
username now defaults to the current users username. -u makes the client
prompt for the username on the prompt.