when hitting an empty line with a single-column load, ascii_io did insert an empty string, first, before quitting the load silently.
IMHO, if it does insert the empty string, there is now reason
to quit on an empty line.
Hence, I removed this "feature", i.e., empty line do not stop
loading anymore.

If quitting is desired, the last empty string should not be inserted
either.

(Multi-column load are not affected, as they issue an error
 on an empty line, comlaining about the missing column seperator...)