roll back martin's latest changes. Its assumption was that implementation routines would rescpect access properties, while all (rightly) assume that a new bat created by BATnew is writable.
alternatives are under study..

(0) not really under study: let BATnew/BATextend return a NULL bat and adapt
all code to check for this and return prematurely with proper garbage
collection.

(1) possibly a setjump/longjump alternative to GDKfatal is possible for
out-of-memory during a BATnew/BATextend. The problem of proper garbage
collection remains.

(2) mguardian does not have garbage collection problems. its drawback is that
concurrent transactions are aborted by a restart.