Integrated SWIG-generated MAPI implementation for Python into the (Unix) build process. Unfortunately, I had to do this outside of the autogen system.  It was just too hard to do inside.
The reason for it being too hard is a combination of peculiarities.
There is a single input file (Mapi.i) which produces two output files
MapiLib.py (name depends on a directive within Mapi.i) and
Mapi_wrap.c.  This is hard (if not impossible) to tell make through
autogen.  Furthermore, the way in which Mapi_wrap.c must be
compiled depends on the Python installation.  This also is hard to
tell make through autogen (by way of a complicated series of configure
tests).

The current setup uses a simple hack in autogen: if there is no
Makefile.ag in a directory, just skip generating makefiles in that
directory.  However, it does generate a list of files-to-be-created by
configure (the .in files).  In the Cimpl directory for Python we now
have a Makefile.in which will thus be copied (with edits) to Makefile
in the build directory.  This Makefile contains all the necessary
commands and uses the standard (for Python) setup.py stuff to build
and install everything.

Also added licenses to the new files.