Merged many of Jan Rittinger's changes into the main branch.
The algebra generation is based on the XQuery Core equivalent of the
input query. Jan has implemented lots of transformations from the
abstract syntax tree to Core in the meantime, that we now need for
the algebra compilation.

I have tagged all the merged files in the Core2MIL_Summer2004 branch
with the tag `merged-into-main-15-07-2004'. To avoid re-merging the
same changes in later merges, please use this tag as the reference
point. (Have a look at the CVS manual on how to do this if you are
unsure.)

Thanks Jan, for that huge chunk of work on the Core compilation!