added running-min/max summaries of [void,X] bats for improving clustered selection
b.rmin(100); # computes a running minimum summary of 100 tuples (sorted on value, reversely on oid)
b.rmax(100); # computes a running maximum summary of 100 tuples (sorted on value *and* oid)

then there are rmin_lower and rmin_higher, to get void-boundaries for range selections
on clustered tables.

That is:
	b.select(lo,hi)

can be rewritten to an oid-restriction, followed by the original selection:

	b.reverse.select(b.rmax(100).rmax_lower(lo),
                         b.rmin(100).rmin_higher(hi)).reverse.select(lo,hi);

this would be especially useful fi we would have in MIL a

		load(b, lo, hi)

such that we could limit the I/O and memory consumption of all column bats that are
reduced by the selection.

In X100 we already have this (extra lo,hi params to BatScan() and Table()).