With my changes from yesterday, compilation of Twig rules failed on all test systems.
This is actually a Twig bug/deficiency. Twig tests its input
file for the suffix `.mt'. This is simply done by searching
the first dot in the input file name, then checking for the
string `.mt'. In the test environment directory names contain
dots, and Twig's suffix checking fails/complains. Using the
original name of the `.mt' file instead of `tmp.mt' triggered
this problem.

As a workaround I've now added `dirname' and `basename'
calls to the Twig command line, eliminating the dots in the
Twig argument.