The W3C formal semantics defines the function fs:convert-operand() to possibly cast the type of an expression e1. The behavior is defined by a second expression argument e2, from which only the type is of interest (not the expression itself).
Many translation rules in the formal semantics specification call
fs:convert-operand() with a fixed type. So far we have thus used
a *type* as the second argument of our fs_convert_op() function in
core.c.

This, however, makes problems with comparison expressions, e.g.

  <a>42</a>/child::text() >= 3

The left hand side evaluates to an untypedAtomic type, the right
hand side to an integer. We thus have to cast the 42 to a number,
while we had to cast it to a string, if a string was on the right
side.

The W3C version does not have that problem, as it applies
fn:convert-operand() to both arguments, with the respective other
argument to define the type to possibly cast to. The semantics of
fn:convert-operand() does the rest.

My changes thus more or less imitate the W3C idea: I've split up
the fs_convert_op() function into two functions:
fs_convert_op_by_type() (which is the function as it was before),
and fs_convert_op_by_expr(), which expects an expression as its
second argument, whose static type is of interest.

It is important to know that the *static* type of the second
argument is sufficient for our purposes. During typechecking we
know the type of fs_convert_op_by_expr()'s second expression and
simply rewrite it to the other variant.

To go into details: I've introduced a new core tree node: stattype,
which is supposed to return the static type of its argument. It
may be used wherever seqtype is allowed. In typecheck.mt.sed,
stattype nodes are rewritten to seqtype nodes with the now-known
type. Subsequent compilation steps will never see the stattype
nodes.