My changes yesterday only considered element construction, but no attributes. The changes now should treat attributes correctly following the W3C semantics.
The fs:item-sequence-to-untypedAtomic() has actually become
obsolete in the handling of attribute values. Instead I've used
the built-in functions fn:string(), fn:join-string(), fn:concat().
(Each of them are F&O functions, so we will have to implement
them anyway. fs:item-sequence-to-untypedAtomic() in turn is only
a formal semantics function, not an XQuery built-in.)
Sorry, Jan, this will probably break your MIL construction code,
but it should be following the formal semantics.

And there's another thing left to do. (I didn't do the change to
not break more of Jan's code ;-)) During element construction, the
W3C demands to
 o merge adjacent text nodes of the element construction argument, and
 o remove empty text nodes.
Both isn't considered so far in the compilation to Core. I suggest
to introduce something like `pf:merge-adjacent-text-nodes()' that
does that task.

While I was at it, I threw out most of the functions in xquery_fo.c
that came in there during my old implementation of a MIL generator
(aka. core2mil.c). They have become obsolete since quite a while.