propagated my recent changes from the main trunk:
In case of errors with individual function calls,
multiplexes use to insert NIL into the result, proceed processing,
and result the result, only issuing a warning that there were local error,
and that NIL had to be inserted instead of the proper result values.

The fact, that the multiplex itself did "succeed" in this cases (as opposed to
failing with an error), made it impossible for the SQL frontend to efficiently
check constraints using multiplexes.

I changed the multiplex implementation as follows:
- in case of errors with individual function calls,
  multiplexes now throw in an error instead of a warning,
- and the multiplex itself also fails in such cases,
  discarding the result.

TODO:
- Should we immediately stop the multiplex processing when the first error occurs?
  For now, multiplex just gathers the errors, but finishes processing.
- What about "hard-coded" multiplexes (e.g. as in modules bat_{arith,math,cast})?
  For now, they do AKAIK only NIL-checking, but no overflow checking?
  Idea: Should they have a global switch (variable) to en-/disable
        nil-checking, overflow checking, etc. (ON = "save", OFF = "fast")?