Fix for bug 1041324  	SQL: weird chars hang server, maxing to 100% CPU
A sloppy implementation of the BlockMode protocol neglected to take care of proper conversions of Strings to bytes and vice versa.
Now, Strings are converted to UTF-8 byte arrays first, before doing measurements on the length of the String.  In some cases (as in the error report) the UTF-8 byte representation is of a different length than the Unicode (Java internal) representation.  On the other side, when retrieving bytes from the server (which are also in UTF-8, since the server 'speaks' UTF-8) a similar error was to be found where the byte-length was taken for String trimming, which could cause an StringIndexOutOfBoundsException.

While fixing this bug, some improvements were made as well, by avoiding method calls and avoiding doing expensive operations (copying strings as in StringBuffer.substring()) when not really necessary.