Undo optimisation, to gain some optimisation Because scrollable resultsets were given a smaller reply_size, this reply_size had to be reset when forward-only resultsets were used.  The JDBC metadata queries use scrollable resultsets, while the default resultset mode is forward-only resultsets.  This results in at least setting the reply_size twice in many cases.
With this change I leave the reply_size setting to the user, who can
suggest the JDBC driver what value to use, and stick to the default
value.  Afterall, it might not be such a big difference for scrollable
resultsets, because it only will differ when the client is just scrolling
back and forth on the edge of two blocks.

Perhaps the block wise approach needs some optimisations in order for
scrollable resultsets to work more efficient in such case.