fixed deadlock in gdk_bbp (decref,incref,BBPlock)
The deadlock was caused by a parallel call to BBPlock (which wants BBPunloadLock
and all swaplocks) and calls to BBPdecref and BBPreclaim which while holding
the swapLock try to get BBPunloadLock.
Fixed by a split of BBP_unload_on in BBP_unload_cnt and
BBP_status_set(BBPUNLOADING). First we set the status of the bat unloading
with the cacheLock and then incrememt the unload_cnt with only the BBPunloadLock