The Windows .rc files contain the DLL version.  This version consists of 4 comma-separated integers, so it can't contain 0_rc1. This checkin implements a smart scheme to overcome this limitation: The SQL vertoo.data file contains 5 parts: - major - minor - release - just a number, 0 for now - rc - the string "rc" for a Release Candidate, "x" for an official   release - build - the build number which doubles as the release candidate   number, so 1 currently, but 2 for the actual Release Candidate.
When a new Release Candidate is being prepared, the build number must
be incremented.

When an official Release (not Candidate) is being prepared, the rc
part must be set to "x", *and* the build number must be incremented.

Setting the rc part can be done with the set-part rc="x" command of
vertoo.py.