fixing BUG in interpret_multiplex_array that was detected by Agustin Schapira:
in a special ("unfortunate") constellation, it may happen,
that a multiplexed binary operator is call with one
[void,any] and one [oid,any] argument of same length,
with the second one actually being dense, and aligned with the
first one, but not being marked dense, and hence not having
a seqbase explicitely stored in the BATheader.
However, since both BAT are detected to be ALIGNsynced
(their halign ids are identical (see below)),
the interpret_multiplex_array
(1) decides (correctly) that the result head can be void,
(2) blindly uses the (not set) hseqbase of the second BAT
    as hseqbase for the result.
Obviously, (2) is incorrect in this case.
Fixed it, by makeing sure that interpret_multiplex_array
propagates the hseqbase of that input BAT that actually
triggers the void result head (because it is indeed marked
dense), and hence has the correct hseqbase set.

Open issue/question:
Did I actually fix a BUG or just cure a symptom?
I.e., is it actually correct, that two BATs can be
ALIGNsynced (i.e., have identical halign ids), with
one being explicitely dense (void), and the other one
being infact oid-dense, but not explicitely marked
as such?
If that latter is wrong, we wuld actually need to
find the operation that creates this inconsistent
situation, but I definitely do not have time for this...

For now, my fix seems to solve Agustins problem, and
that is my only concern right now.


Unfortunately, I do not have time, now, to also provide the
respective test script to confirm that the BUG it indeed fixed,
and to prevent re-occurance of the same bug.
The script that triggered that bug (provided by Agustin) requires
a special database to create the special ("unfortunate") constellation,
and I have no time to adapt the script and database to suite our
testing environment...