BUGFIX [ 1108971 ] Syntax errors crash Mserver
fix: do not use REJECT anymore (apparently input()/unput() usage inside
lexer actions makes it impossible to use REJECT).

REJECT was used to back off from a sequence of chars that are permitted
as operator names; if that sequence did not match any existing operator
in a loaded MIL module.

instead; the sequence of operator-chars that do not match an operator
is given back to the lexical analyzer buffer with unput()  *except* the
first character, which is emitted as a single char token.

REJECT was not needed as the char sequence could not match any other
subsequent rule anyway.