Fixed up the cleanup code after a module has been loaded. When a module is loaded, the file is read into allocated memory which is supposed to get freed once the module has been parsed.  This didn't quite happen this way.
Also, the variable monet_script contains a list of pointers to scripts
on the command line.  The first time scheduleClient is called, these
scripts are executed.  The list is allocated in main but is never
freed.  We now free the list after the scripts were executed, and we
can now also use the the value of the pointer to see whether this was
the first time scheduleClient is called, so we don't need the static
variable monet_first anymore.