Check whether the database is binary compatible.
This is done by adding a line to the top of the BBP which contains two
numbers: sizeof(size_t) and sizeof(oid).  If such a line is found, but
the numbers don't match, Mserver will exit immediately with an error
message.

If the line is missing, there are two alternative implementations.
The currently selected one (CONTINUE_IF_NO_SIZES_IN_BBP defined) is
that Mserver will issue a warning message and assume the database is
compatible (if it isn't, and you don't quit(), you'll most likely get
a crash and a corrput database).  The other implementation will issue
an error and exit.

The symbol CONTINUE_IF_NO_SIZES_IN_BBP is defined in this file.

This fixes SF bug #1108525.