Two changes: - When matching strings (single or double quoted), have (f)lex match   only the opening quote and do the reading of the rest of the string   ourselves. - Remove all occurrences of REJECT.
The reason for removing all REJECTs is so that input() will work on
(very) long strings.

The way in which REJECTs were removed was as follows.  REJECT was only
still used to reject matches of negative constants when the preceding
token was an identifier of constant.  In that case, the negative
constant was really a binary operator "-" followed by a positive
constant.  We now set a start condition called "idconst" when we
encounter an identifier or constant and only match the negative number
if the start condition is INITIAL (aka 0).  So we also set the start
condition to 0 on all other matches (the BEGIN(0)).

This fixes SF bug #1120061.