The previous fix still didn't fix the problem of a hanging tcpip1/tcpip2 test on Windows.  Now fixed it the other way round.
The problem seems to be that if you're using the tcpip module to
communicate between two threads in the same Mserver (which is what
those tests do), Mserver will hang.

The way we're trying to fix this is by not having two listeners on the
same end of a socket.  The first attempt tried to not start a listener
by the "receiver" thread, and this attempt tries to not start a
listener by the "sender" thread.  And of course, the threads should
get started if multiple Mserver processes are involved, whether on the
same machine or on different machines.

In the fix we use the ports array which is used to store all ports for
which there are listeners in the current server.

Also fixed a potential array bounds error filling the ports array.