And here is the final (?) part 2 of my loop-lifted-staircasejoin optimization.
Basically, I stipped-down the C implementation of the loop-lifted-staircasejoin
(ll_staircasejoin.mx) to the very core of the child/desc/desc-or-self steps,
assuming that the input is sorted on item+iter and the result also has to be
sorted on item+iter. This makes the "bucket" feature of the desc/desc-or-self
implementation obsolete. The desc/desc-or-self implementation handles the
special case that there is only one identical item for all iter; in that
case, the desc/desc-or-self is just a simple slice of the candidates list,
replicated for all iters.

The C implementations "PFll_(child|descendant|descendant_of_self)" are
called from a MIL wrapper that takes care of expliting the given order
properties (iter+item or item+iter) of the input, and generating the
requested order properties for the output.

Furthermore, in function "loop_liftedTextConstr" of "milprint_summer.c",
I improved the part "/* get the property values of the strings; */": we now
invest in sorting the two BATs that are to be joined on a string column. It
turned out, that (at least for large BATs) the merge-join on strings
(including the sorts) is faster and (performance-wise) more robust than a
hash-join. This improvement makes q10 sclae linearly up to 10GB --- at least
on the 16GB machine.

A comparison between the previous version ("u") and the latest version ("8")
reveals that the above changes are a small "investment" on small docs <= 1MB
that "pays off" with large docs (>= 100MB):

(data from the "alternative" 16GB machine --- results for the "regular" 8GB
machine will follow)

     116KB  116KB     1MB    1MB    10MB   10MB   100MB  100MB       1GB      1GB      10GB     10GB
         u      8       u      8       u      8       u      8         u        8         u        8
01  0.040u 0.0508  0.042u 0.0528  0.054u 0.0648  0.197u 0.1968    1.647u   1.5928   16.4430  15.6268  01
02  0.049u 0.0638  0.052u 0.0678  0.089u 0.0908  0.446u 0.3658    3.552u   3.0098   34.9880  29.4548  02
03  0.146u 0.1968  0.167u 0.2128  0.341u 0.3578  1.933u 1.7838   16.523u  15.2438  163.7210 146.9678  03
04  0.078u 0.0998  0.083u 0.1018  0.102u 0.1118  0.590u 0.5268    5.214u   4.6578   48.9690  43.0578  04
05  0.040u 0.0558  0.042u 0.0568  0.064u 0.0688  0.209u 0.1948    1.799u   1.4658   18.2840  14.6218  05
06  0.020u 0.0278  0.021u 0.0278  0.025u 0.0318  0.059u 0.0708    0.398u   0.4128    3.8880   3.8628  06
07  0.024u 0.0338  0.025u 0.0338  0.031u 0.0418  0.080u 0.0938    0.568u   0.6458    5.8110   5.7558  07
08  0.088u 0.1158  0.099u 0.1228  0.175u 0.1788  0.988u 0.9168   14.335u  13.4868  315.3940 304.5438  08
09  0.133u 0.1838  0.145u 0.1928  0.237u 0.2558  1.140u 1.0698   17.197u  16.2448  406.9060 394.7218  09
10  0.327u 0.4588  0.411u 0.5108  0.970u 0.9708  6.784u 6.2808  107.478u  63.4188           665.5878  10
11  0.093u 0.1298  0.106u 0.1368  0.233u 0.2318  4.597u 4.4478  430.724u 429.0648                     11
12  0.104u 0.1468  0.113u 0.1478  0.179u 0.2158  2.167u 2.0808  199.107u 197.9108                     12
13  0.054u 0.0758  0.056u 0.0778  0.078u 0.1098  0.243u 0.2448    1.893u   1.7768   19.1860  17.2418  13
14  0.036u 0.0508  0.061u 0.0738  0.327u 0.3138  2.745u 2.6948   27.469u  26.7018  939.4790 907.9828  14
15  0.051u 0.0778  0.068u 0.0948  0.093u 0.1158  0.305u 0.3088    2.607u   2.2878   25.2620  22.3978  15
16  0.059u 0.0858  0.075u 0.1188  0.103u 0.1318  0.351u 0.3498    2.975u   2.5058   28.6940  24.5458  16
17  0.050u 0.0688  0.054u 0.0778  0.092u 0.0978  0.458u 0.3768    4.184u   3.5898   40.4330  33.6498  17
18  0.033u 0.0458  0.034u 0.0558  0.054u 0.0588  0.186u 0.1958    1.417u   1.2448   15.0050  12.0698  18
19  0.066u 0.0948  0.076u 0.0998  0.160u 0.1628  0.845u 0.6928    8.098u   6.7868  123.3520 107.5638  19
20  0.196u 0.2448  0.203u 0.2838  0.254u 0.3008  0.861u 0.8068    6.854u   6.0748   68.9980  59.7138  20
21  0.021u 0.0248  0.064u 0.0718  0.482u 0.5168  4.525u 4.7838   45.166u  47.4608                     21
         u      8       u      8       u      8       u      8         u        8         u        8
     116KB  116KB     1MB    1MB    10MB   10MB   100MB  100MB       1GB      1GB      10GB     10GB

(On the 16GB machine,)
all queries but 8, 9, 11, 12, 14, 19, upd scale linaearly to 10GB.
 + 8 & 9 "suffer" from the theta join(s) but stay with a factor 20.
 + 21 runs out of memory/disk space
 + 19 suffers from a CTrefine that most probably runs out of memory
      (but stays within a factor 16)
 + 14 seems to run out of memory with several operations
      (a "string_join" be the most expensive one...)

11 & 12 obviously don't fit...
... though there woudl be a way to make them fit:
in both cases, the result of the (only) expensive theta join
is only used for a {count}; hence, we could "pipeline" the
theta-join plus the {count} by replacing
	var res := theta-join(l,r);
	var aggr := {count}(res);
by (some thing like)
	var x := kunique(l);
	var aggr := bat(htype(l),int,count(x));
	x@batloop() {
		var lx := l.reverse().select($h).reverse();
		var res := theta-join(lx,r);
		aggr.insert($h, count(res));
	}
thus, we would at least avoid the huge intermediate result
of the theta-join.
Open questions:
 - can PF recognise the fact, that a join result is only use for a {count}?
 - should we look into this before wednesday?


PS:
the changes also fix tests
    tests/XPath/Tests/q04.xq
    tests/XPathMark/Tests/Q4.xq

only test
    tests/XQuery/Tests/step_following-sibling1.xq
still failes and needs to be checked later...