Ported to MinGW.
This version compiles using the MinGW suite, but doesn't necessarily
work correctly.  It shouldn't have any deleterious effects on other
platforms.

The one big user-visible change is the renaming of the "kernel" module
to "builtin".  For some reason it was impossible (with the limited
amount of time available to search for solutions) to get the dynamic
loading of the kernel module to work.  We know it's not the code in
the module since renaming the module worked, and deleting all the code
didn't work.  This change shouldn't affect users too much, though,
since the module is loaded automatically at startup and is probably
never loaded explicitly (except in testing).

Other changes are mostly parameterizing the makefiles (we use the
msc makefiles), changing or adding ifdefs.

I also moved the code for handle_argerror and handle_paramerror from
the builtin (kernel) module to monet_interpreter.mx so that mapi.mx
doesn't need to link with builtin (which didn't work on MinGW).

Finally, Mserver.mx does not use C++ anymore.  It was not used.  This
change is not essential and can be reverted by changing all occurances
of "@c" to "@C" in Mserver.mx.

Test output has been updated appropriately.