Removed the batHole. Beware this again changes the bat descriptor, ie you need to save/restore your db. The hole needed to go as we are about to implement an api for updating blocks on a block granularity instead of full bats.
With the hole gone we have again a value at bun postition 0. This means
the hash chain lists now end with HASH_MAX instead of 0.

Also MonetDB is ported to WINE/mingw32. Still 34 tests fail
mostly because the bats are allmost allways unloaded (because the
amount of internal memory isn't found out correctly) and winsock on
wine is as buggy as on native windows.