This is a modified fix for bug #1207124.
When optimizing for clauses that bind to just one item, don't
forget to set types correctly.

In contrast to Jan's fix, this one does only set types for
those nodes that are actually affected. Jan's fix re-checked
the entire subtree which is correct, but does unneccessary
work that might hit performance.