Forwarding Sjoerd's fix for documentation purposes.
This patch avoids the "pf_compile_interface: unknown output mode"
message when using mTest.  mTest uses a 'ping' to check whether the
server is yet up, this is done by sending a single line consisting of
the word "ping".  Due to my changes CMDxquery is now called with an
empty string which causes the above warning message.  This message is
now sent to the client (where the warning belongs) and since any output
on the 'ping' command issued by mTest, the warning disappears.