Fixing a bug that appeared on Darwin as: "Interval field seperator ':' missing"
This bug has actually nothing to do with Darwin, but with the time zone
the machine is located in.  The statement emitted was:
"SET TIME ZONE INTERVAL '--08:00' HOUR TO MINUTE;"
which caused the above error message.  Although the error message is a
bit misleading, it indeed points out that the given string is not valid.

Fixed it by suppressing the minus-sign and letting the
Integer.toString() method place it for us.