Checking in on behalf of Jan Flokstra:
I already said this Mapi stuff was a major headache for me when
concerning escapes.  The fix for this problem which works with me is to
add an extra condition on line 553 in runtime/serialize.mx (the line
above start--; ....

&& (src[end]!='\n')

This ensures the extra '='is printed in Mapi mode.  Hope this works.