Fixing a nasty bug in the serializer that cost me a couple of hours and Nina even more of her precious time ... |-((
The details:
~~~~~~~~~~~~

Shredding a document that contained some "strange" characters like

        "Ã"

in MonetDB/XQuery worked fine, but when these characters occurred in a
query's result, Mserver crashed with a segfault:

========
$ echo '<NAME VALUE="Ã"/>' | pf | Mserver --set monet_prompt='' ; echo -e "\n$?"
--------
# Monet Database Server V4.8.2
# Copyright (c) 1993-2005, CWI. All rights reserved.
# Compiled for x86_64-redhat-linux-gnu/64bit with 32bit OIDs; dynamically linked.
# Visit http://monetdb.cwi.nl/ for further information.
<?xml version="1.0" encoding="utf-8"?>
<XQueryResult>
<NAME VALUE="Ã
1
========

Debugging revealed, that the segfault actually occurred in function
"emit_escaped()" of runtime/serialize.mx, more specifically in line 552:

        char *emit_val = isspecial_val[mode+(int)isspecial_ch];

Apparently the calculated index (mode+(int)isspecial_ch) was larger than the
size of array "isspecial_val[]". The following log depicts the details
(a temporary patch avoids the segfault):

========
$ echo '<NAME VALUE="Ã"/>' | pf | Mserver --set monet_prompt=''
--------
# Monet Database Server V4.8.2
# Copyright (c) 1993-2005, CWI. All rights reserved.
# Compiled for x86_64-redhat-linux-gnu/64bit with 32bit OIDs; dynamically linked.
# Visit http://monetdb.cwi.nl/ for further information.
<?xml version="1.0" encoding="utf-8"?>
<XQueryResult>
! runtime/serialize.mx: emit_escaped(): mode=1, end=1,                         (int)src[end]=-115, isspecial_ch=isspecial_tab[               (int)src[end]]=
! runtime/serialize.mx: emit_escaped(): mode=1, end=1,                (unsigned int)src[end]=4294967181 >= UCHAR_MAX+1=256 !
! runtime/serialize.mx: emit_escaped(): mode=1, end=1, (unsigned int)(unsigned char)src[end]=141        <  UCHAR_MAX+1=256 !
! runtime/serialize.mx: emit_escaped(): mode=1, end=1,          (int)(unsigned char)src[end]=141,  isspecial_ch=isspecial_tab[(int)(unsigned char)src[end]]=
<NAME VALUE="Ã
! runtime/serialize.mx: emit_escaped(): mode+(int)isspecial_ch >= 32 !
"/>
</XQueryResult>
========

Due to "char *src", "src[end]" can be negative (surprise!??), and hence
"(int)src[end]" will be negative as well, which is not a good idea for an
array index ... |-(

Although line 549 of runtime/serialize.mx seemed to be aware of this:

        stream_printf(GDKout,"\\%03o",(unsigned char)src[end]);
                                      ~~~~~~~~~~~~~~~

line 544 was not:

        if ( (isspecial_ch = isspecial_tab[(int)src[end]]) ) {
                                           ~~~~~

patching the latter as follows seems to solve the problem:

        if ( (isspecial_ch = isspecial_tab[(int)(unsigned char)src[end]]) ) {
                                                ^^^^^^^^^^^^^^^


To be honest, I do not understand exactly, how the whole code in function
"emit_escaped()", using the arrays "isspecial_tab[]" and "isspecial_val[]",
is supposed to work.
Hence, I have no idea, whether my fix is actually correct or not ...

Thus, could some one who does understand this code (Jan F.?) *please* check
my fix, and the whole code, and verify that both are correct and
"water-proof", now?

(Unless I'm the only one who's too stupid to understand this code easily,)
Maybe, there's also a way to make this code more ready, and hence
more maintainable and less error-prone...?

[
ps:
Even if the "evil" character is "invalid", MonetDB/XQuery should not crash
with a segfault, but rather issue a proper error message, preferably already
during shredding, or while parsing the query ...
]