Also with gcc 4.0.1 on Fedora Core 4 on x86_64, the XQuery compiler crashes with a segfault, mostly in a deeper recursion of function reduce in line 675 of revision 1.19/1.17.4.2 of file compiler/core/fs.brg .
The same crashes seem to occur with Intel's icc 8.1 on IA64 (Itanium).

Note, in both cases, the crashes only occur, if the XQuery compiler is called
within Mserver (i.e., via "xquery(<mode>,<query>);" or via "MapiClient -lxquery"),
but not when the stand-alone compiler "pf" is used.

See
http://monetdb.cwi.nl/testing/projects/monetdb/Stable/pathfinder/.mTests103/index_short.html
for details.

I suspect that problem are related to BURG-generated code and/or the use of longjmp()
(or the combination of both), which seem to trigger some bug(?) in these compilers.

Note, with gcc 4.0.1, the number of failing tests seems to depend on the optimization level:
 - configure [default]         (=>  gcc -g -O2 ):  9 tests fail
 - configure --enable-debug    (=>  gcc -g -O0 ): 45 tests fail
 - configure --enable-optimize (=>  gcc -O6 ...): 42 tests fail

For gcc 4.0.1 on Fedora Core 4 on x86_64, we avoid the problems
by falling back to gcc 3.2.3 ("gcc32")...