fixing BUG [ 1250067 ] MIL: time() limited to < (+-4000s)
The problem is actually not "time()" (which returns milliseconds
as int, and hence allows to cover 2147483 secs ~= 24 days),
but "usec()", which returned microseconds as int, and hence allowed
to cover only 2147 secs ~= 35 minutes.
Since the underlying GDKusec provides the milliseconds as lng,
there is no reason why "usec()" should not return a lng as well.