Fix for bug 1274903:
Specific typing rule for pf:item-sequence-to-node-sequence() did
not correctly handle the type `item'. The typing rule implemented
in PFty_is2ns() now makes sure that it always returns a subtype
of item*.

Given that we have a specific typing rule for that function it is
not sensible to overload it in xquery_fo.c. -> removed overloading.

Added test for the situation.