THIS IS THE MONETDB/XQUERY UPDATE BRANCH. This branch is to be used in conjunction with the XQuery_0-9_update branch.
Added new versions of the various replace functions that take an extra
argument (bit force) to force replacement, even if the BAT was
read-only.  It also does the replacement inline (as opposed to first
deleting the old and then appending the new) so that void-headed BATs
do not become oid-headed.