... more "print_error - tuning":
For MAPI, we need to prefix all error messaged with a "!",
as they (might) get lost otherwise.
Hence, I added a hack(?) that (mis?)uses the ctx parameter
of the error handler to remember whether a newline occured
across multiple calls of the error handler for one error
message, and hence allows us to prefix each (line of an)
error message with "!", or better with the usual "!ERROR: ".