- undo hack to allow empty statements - allow module declarations
when a module declaration xquery script is executed, it produces
the empty sequence.

module import is not yet supported, however, I'm thinking about
a simple implementation that assumes the code of module NAME in
namespace URI to be present at URI/NAME.xq (unless specificed
otherwise with "at"). It then just becomes a hack in the scanner
to fetch this URI and keep parsing (though inside a module everything
but let-statements and methods should be ignored, and module import
loops should be detected and lead to errors unless all are inside
the same namespace).

This proposed implementation is back-end agnostic, thus future-proof.
I could use some help with it, Jens.

However, for MonetDB/XQuery, I then want a shortcut in the query
handler that before entering the real PF parser detects queries that consist
of a single module import and a method call. These can be handled using
a MIl tree query cache (similar to the cache in SQL). That one will reduce
the MIL overhead in the Philips benchmark, and is also just what we need
now for handling soap calls.