Added a shortcut (HACK) to improve SORTfndfirst & SORTfndlast on BATS with all tail values being equal: In that case, the first binary search step did hit right in the middle, and then each, SORTfndfirst & SORTfndlast, did linearly scan one half of the BAT to find the first/last matching BUN.
Instead, we now first check, whether the first/last BUN is larger/smaller
or equal to the search value, and if so, immediately return
BUNfirst/BUNlast.

Also fixed the documentation of SORTfndfirst & SORTfndlast to match the code:
"
We have two main routines, @`SORTfndfirst@5(b,v) and @`SORTfndlast@5(b,v), that
search for a TAIL value v in a sorted BAT. If the value is present,
the first routine returns a pointer to its first occurrence, while the second
routine returns a pointer to the BUN just after the last occurrence of v.
In case value v does not occur in the tail of BAT b, both routines return
a pointer to the first BUN with a tail value larger than v
(i.e., @%BUNfirst(b)@, in case all tail values are larger than v);
or @%BUNlast(b)@, in case all tail values are smaller than v.
"