Implemented updates.
This version is an intermediate step towards the real thing and is
intended as a proof-of-concept and an indication of what needs to be
changed in the documents being updated.

Each of the operations first create a "replacements" BAT which is a
BAT of BATs with the head value the BAT that is to be changed (i.e.,
this BAT is *not* changed during the operation), and the tail the new
values (either replacing existing values or new values).  This
replacements BAT is executed at the end of the operation.  This is
something that will need to be changed: the replacements BAT should be
made at the start of the query and executed at the end.

One essential change that will need to be made to the pathfinder
machinery is that currently the "item" BAT contains PRE values.  This
needs to be changed, at least for all update operations, to NID
values.

I added a bunch of supporting PROCs to pf_support.  A more generally
useful one is a version of mposjoin where the frag parameter is an OID
instead of a (fake) BAT.

The bulk of the update code is also in pf_support in the form of
several PROCs.