Implemented updates using the Update Tape.
This code is more-or-less completely untested, but it compiles and you
can import the modules.  (In fact, I did one (yes, one) test and that
seemed to produce the correct result.)

One open question (for me) is: how can I recognize that a query is an
update query?  I currently use an extra variable is_update which is
initialized to false and set to true in the code that implements the
various update functions.  There might be a better way.

The code is intended to work with the PRE values in the input item
BATs and there is no need to change those BATs to contain the NID
values.