Though Todd Proebsting's BURG code has proven amazingly solid, in fact I found a bug in there over the weekend.
Burg so far did not correctly handle percent signs `%' that
appear in the prolog of a burg file.  This works fine as long
as you don't want to do any modulo operations in the prolog.

Fixed burg's lexer accordingly.  It now really just looks out
for `%}' (a percent sign followed by a curly brace) that
terminates the prolog.