The second batch...
This checkin enables the Burkowski steps, so that they can be used in
the XQuery language. /select-narrow:: /select-wide:: /reject-narrow::
and /reject-wide::

There has been some opposition against tampering with the XQuery language
as set by the W3C. I am aware of that and I am not trying to change
the language. The code added by this checkin will only add the steps
if the source code is compiled with the "--enable-burk" flag, which
is turned off by default.
I marked all the code added by this checkin with either an #ifdef
BURKOWSKI or with /* [BURKOWSKI] */ -markers. I made tags just before
and after the checkin as well (pfBurkowskiBefore, pfBurkowskiAfter).

The reason for checking in these steps, is for maintenance reasons and
for distribution (there are several persons/groups that do want to use
these new steps).

None of the steps have been thoroughly tested and there will certainly
be a lot of work to do. The next checkins (in this series) will add
test-scripts and will fix some open issues.