Commit the SOAP RPC extension of the pathfinder compiler:
In the struct 'PFcsem_t', the record:
    struct PFfun_t *fun;   /**< function reference */
is replaced:
        PFapply_t  apply;   /**< function application */

The struct 'PFapply_t' holds the information in a PF parse and
core tree for function application. In the past, this information only
consisted of the function pointer. Now, we remember whether it was an
rpc or not.  (see also Peter's commit of 'apply.h').

This change results mainly in replacing the *sem.fun* in the code with
*sem.apply.fun*.