tentative fix for bug 1405188 (the document shredding issue) - pathfinder.mx: docid was passed as name instead of number - shredder.mx: prevent empty [void,void] attr_prop bat from being created                (such bats later cause mposjoin errors when mingled with                 non-empty [void,oid] bats)
The fix is TENTATIVE as it is not checked into the HEAD, rather in
my current work env (the update branch).

I'd suggest apllying it to the HEAD, and then checking the test web.
My current on-the-road setup (MinGW) cannot run the test web. Thus
I cannot guarantee correctness (a show-stopper just before the release).

However, the bug is quite serious so I would recommend including the fix
in the release. If nobody steps in I will do so when I am back in .nl