Backport fix:
Fix for bug [ 1211372 ] PF: Mserver crashes on space and/or '\' in
shred_doc path.

The problem was, the error message generated by xmllib contains the
%-escaped (url-ified) file name.  So, "Documents and Settings" becomes
"Documents%20and%20Settings".  This string is given unaltered to
stream_printf and is interpreted as format string.  %20a and %20S
happen to mean something (look it up).