Fix: when appending a BAT to an empty [void,oid] BAT with tdense set, the tdense flag was not correctly set afterward.  We now copy the denseness information from the appended BAT.
Also, if we're appending a void-tailed BAT with seqbase set to an
empty BAT, we set the seqbase of the empty BAT to the seqbase of the
appended BAT so that we don't materialize unnecessarily.

Finally, to compensate for all the extra checks, an unnecessary
condition on sz not being zero was removed (if sz==0 we already
returned in at the start of the function).