Move the MIL code, which retrieves results of SOAP-RPC calls out of the working set and builds the iter|item|kind BATs, from milprint_summer.c into a new PROC 'get_rpc_res' in pf_support.mx.
The main reason of this action is because "batloop" is used in this
piece of MIL code.  Since the simplifier of MIL code does not recognize
loops, it incorrectly removes code which is really necessary.  By moving
this piece of MIL code into a PROC, we can prevent it being simplified.