last-minute name-change.
Some people (including myself) thought that the initial name "Burkowski Axis steps"
was not completely correct, due to a difference in semantics between the original
ideas of Forbes Burkowski and the axes implemented here.

While it is still possible (looks like this is the last and only chance there is),
I'm renaming the steps to StandOff Axis steps, which seems to be more appropriate.

Every reference to Burkowski Axis steps should have been removed with this check-in.