improvement to the heuristics for allocting VM memory wisely:
in VM, we allocate pages, some not yet committed, but the VM rage
already reserved. advantage is that the heap can be allocated
without risk of base-reallocation (this may or may not be
accompanied by a copy step - OS implementation dependent).

the heuristics try allocate extra reserved space if a base
reallocation cannot be avoided anyway. If it can be avoided
(the memory need is within the already reserved area) no
extra space should be reserved.

regrettably this is done in two places:
  * in BATextend / BATgrows it was done correctly.

  * in HEAPextend it is done again. I think this is because
    the HEAP interface caters for more than BUN heaps only.
    However, no attention was paid whether we are actually extending
    beyond the maxisze. If this is not the case and a base reallocation
    is unnecessary, we should *not* allocate extra reversed space.