Fix a bug in the RPC receiver: currently, if an RPC request a XML node parameter, the XML node is represented in a wrong way, namely it is treated as a part of the request SOAP envelope.  The problem with this representation is that if one of the steps (all 'ReverseAxis' steps plus two 'ForwardAxis' steps):         ancestor         ancestor-or-self         parent         preceding         preceding-sibling
        following
        following-sibling

is queried on the XML node, wrong result will be returned.

For example, the 'paren' step will return '<xs:anyNode>' which is the
parent node of the XML node in the request SOAP envelope, and the
'ancestor' step will return the whole SOAP envelope.

What we need is to partition the XML nodes in a request envelope into
fragments, something very similar with the transient documents in the
working set.  But WS_FRAG currently only exists for transient
documents, so we need to extend it so that there is a WS_FRAG BAt
available for every document.

We have changed the current WS_FRAG bat from a [void,oid] BAT into a
[void,bat] BAT which in turn contains [void,oid] BATs.  This way, we
have a bat for every document to denote fragment boundaries if
necessary, although it is currently only used by the transient
document and the document containing an RPC request message.

The creation of the WS_FRAG BATs is moved into CMDshred2bats.
CMDshred2bats just creates an empty WS_FRAG BAT for every shredded
document.  This approach makes it easy and fast to add the WS_FRAG BAT
into the working set, namely, in 'index_doc' the temporary 'shredBAT'
is added to the working set by a multiplex 'insert' operator, and
since the WS_FRAG BAT is contained in 'shredBAT', it is automatically
added.

The changed code uses an 'if' statement,

    if (ws.fetch(WS_FRAG).fetch(frag).count() > 0)

to see if WS_FRAG BAT should be used.  Note that the following check
won't work:

    if (exist(ws.fetch(WS_FRG), frag))

since a WS_FRAG exist for every document.

The final note: this change requires that all previous shredded
documents need to be re-shredded!