The semantics of (statically typed) empty sequences can be quite hairy.  In the query
  for $v in () return e ,

for example, we cannot assign a sound type to variable $v.  In line
with the XQuery specs (see FS, beginning of Section 4), this checkin
prohibits the use of the empty sequence in the `in' part of a FLWOR
clause.  (The result type of such a FLWOR would be the `empty', which
violates the requirements of the specs.)  Two queries in the TestWeb
were actually illegal for this reason, and I adapted the stable
output accordingly.

On the back-hand side, this restriction breaks things in some
situations that _are_ legal in XQuery.  Our normalization phase
(fs.brg) might normalize legal queries into illegal ones.  The
simple query

  sum (()) ,

for example suffered from this problem (and actually made us realize
it).  To cope with this, I modified the normalization to _not_ introduce
illegal iterations over literal empty sequences in function applications.

My checkin does not make us fully compliant to the specs with respect
to (statically typed) empty sequences, but we're at least getting
closer...