Added a 'modules' subdirectory to the pathinfinder root to be used for XQuery modules in Pathfinder which are not 'standard' XQuery.
I defined a module 'pftijah' in the modules directory with at the
moment only a dummy module to test if my attempt to conditional
compilation works. The module is only compiled with a have_pftijah
configuration switch eg.

./configure ..............  have_pftijah=yes

The main reason for this is to ensure that regular Pathfinder developpers
and users will not have any compilation/runtime overhead of our work. The
only overhead is the fact that the software is in the CVS tree. We
have tried to develop a version outside the Pathfinder CVS tree but we
ran into all kinds of software engineering and synchronization problems.