Almost a complete rewrite of the Update code. The Update tests ran once without error, so it's time to check this in.
We distinguish two kinds of columns in the working set: PRE/RID
columns and the rest.  The PRE/RID columns are divided into pages of
2**16 tuples.  Other columns are not divided into pages.  We may
insert new PRE/RID pages, but we don't insert pages into any other
columns.

When inserting into a document, we check whether there is enough free
space (holes) on the page where the insert is to take place (anywhere
on the page, we'll move data if needed).  If this is the case, we use
the old method of creating "updates".  If there is not enough space to
be found on the page, we create new pages and insert them after the
page on which the insert is to take place.  The data after the insert
point is moved to the end of the last inserted page, and the
to-be-inserted data is copied in.

New pages are logically writable (and hence written to during the
update -- logically means that the BATs are read-only but we can use
the "force" flag to ignore this fact), existing pages are not written
to during the update phase: they are only written during the commit
phase which comes after all updates in the query have been done.

For all other updates, we never create new pages.  However, if a new
page exists, we write to it, otherwise we maintain a list of updates.
Because of the order in which updates take place (first non-structural
ones, then inserts, and finally deletes), this is only an issue for
insert and delete (i.e. for non-structural updates we maintain the
update list).