(Commiting Peter's fix)
Start 'Mserver' with the module 'xquery_rpc', then call
'rpc_receiver_start' multiple times with different port numbers will
result in error like:

!ERROR: interpret: unknown variable '72.thash'.
!ERROR: interpret_params: httpd_start(param 1): evaluation error.

Have a look at in the parse tree, we can see that the value of the first
parameter of 'httpd_start', which is the port number, is overwritten
unexpectably.

This problem seems to be caused by calling 'yycopy' with kill_orig = 1
in Myycopy and it seems to be fixed by calling 'yycopy' with
kill_orig = 0.