peter discovered a new field to use the StandOff steps for (bio-informatics, analysing chromosomes).
Appearantly the XML documents to be analysed have, unlike text-documents, many
overlapping regions. In the current implementation a stack overflow would occur
in the loop-lifted steps.
To fix this bug, this patch will make the size of the stack increase automatically when its
maximum is reached. Because the steps are optimized for a small stack, a warning
will be issued whenever the stack-size needs to be increased.