Fix for bug 1499957.
When doing rcopy(), let the first and inserted pointers in the copy
point equally far into the bun heap as the source bat.  If this isn't
done, for some operations (BATloop and BATpropcheck come to mind) it
looks like there are more BUNs in the copy than in the original.