curing (rather than fixing...) the basic/test2 part of bug #1509413 "PF: various StandOff tests fail":
When attr_own is empty, and the tail type is physically 'void',
it 'sometimes' happens that the tail's seqbase is not set (i.e., 'nil'),
which makes the tail logically 'NIL' iso. 'VID' or 'OID',
and hence causes the insert to fail due to tail-type mismatch.
Since I haven't found the place where the empty attr_own 'looses' its tail seqbase,
and since neither the tail-seqbase nor the insert has any impact if attr_own is empty,
we simply avoid the problem by skipping the 'NOP'-insert completely in that case.