Porting the radix module from M4 to M5 --- an educative story.
This is Phase 1.
See details below.


Probably a couple of years too late, but here's a (not that) short story
that describes the adventure of porting a MonetDB version 4 (aka. "M4")
MEL-/MIL-based extension module into a MonetDB version 5 (aka. "M5")
MAL-based extension module, with the major goal to retain as much
functionality as possible while keeping the differences between the
resulting code bases as small as possible to allow for (semi-)automatic
propagation of future bug fixes, feature extensions an other code changes in
both directions as long as both versions of MonetDB are still active.

Migrating as much as possible of the mature code base of version 4 (as well
as future bug fixes) is a prerequisite to develop version 5 with the obvious
and precious head-start as opposed to once again going through more than 8
years of code base consolidation and stabilization --- at least for those
part that version 5 inherits from version 4.

To meet these targets, migration of an extension module from version 4 to
version 5 requires two major phases.


Phase 1 happens only in the version 4 code base, and is mainly geared at
preparing the code thus that a subsequent porting to version 5 requires only
minimal code changes. This does not only allow semi-automatic (and hence
more efficient and less error-prone) propagation of code changes between the
to versions, but it also clearly separates version-independent code
changes/improvements from those changes that are particularly required for
the migration. Again, this "planning" helps to limit the risk of bug
sneaking in unnoticed.
In principle, Phase 1 needs to fulfill two major tasks.

1a) "de-MIL-ify" the code.

    Next to providing MIL command that are implemented in C, most M4
    extension modules also provide some MIL code, mainly defining extra MIL
    functions ("PROCs").
    Obviously, M5 does not come with/support MIL anymore. Moreover, there is
    no comparable MAL procedure/function definition possible in M5/MAL
    extension modules.
    Hence, "de-MIL-ifying" means to replace these MIL PROCs by MIL commands
    that are implemented in C in order to retain their functionality in/for
    M5.
    "de-MIL-ifying" does not "harm" the M4 code base. In contrary, recent
    analysis has shown the MIL PROCs are both space and performance wise
    less efficient that MIL commands implemented in C. Hence, next to
    preparing a "minimal-invasive" migration to M5, "de-MIL-ifying" also
    might improve the M4 code base.

1b) Cluster version-dependent and version-independent code.
    Again, the goal is to keep the final difference between the to code
    versions minimal, by avoiding/removing any interleaving between code
    that can/will remain unchanged during the migration and code that need
    to be modified.
    With the MIL code by task (a), this boils down to clearly separating the
    MEL code (@m), the header file code (@h) and the C implementation code
    (@c).

1c) In addition, I took the opportunity to distinguish exported from only
    locally use functions by prefixing the former with
    		"M4_RDX_"
    and explicitly declaring the latter as
    		"static".


Phase 2 then happens only in the version 5 code base, after Phase 1 is
finished on the version 4 code base. In Phase 2, the whole code it first
copied to the M5 repository and then ported the comply with the M5
environment. Since M5 does not have/use MEL any more, the major task is to
basically perform MEL "by hand" in order to create the interface between the
C implementation and the MAL interpreter.
Here, there are four tasks.

2a) MEL (@m) -> MAL (@mal).
    First, the MEL definitions of the command signatures (@m) need to be
    translated in to the respective MAL syntax (@mal). While most things are
    straight-forward (provided one is familiar with both the MEL and MAL
    syntax and semantics, there is one minor but nasty issue that I consider
    worth mentioning.
    MAL does not (seem to) allow whitespaces around ':' that separates an
    identifier from its type. While this "feels natural" for variable and
    parameter/argument definitions (i:int, b:bat[:oid,:lng]), its not that
    "natural" that MAL force one to append the return type of
    functions/commands without whitespace to the respective signature, i.e.,
    if you liked to write, e.g.,
    		.COMMAND foo ( int bar ) :
    			 lng = MyFooBar;
    in MEL, you now need to write
    		command foo( bar:int ):lng
    		address MyFooBar
    in MAL.

2b) Adding wrapper functions.
    The calling conventions for the C functions the implement MIL/MAL
    command have changes between M4 and M5. Both (still) use the first
    parameter to return the actual result and use (typed-)pointers ("call by
    reference" to hand the remaining arguments to the function.
    However, while M4/MEL expected the functions to return an int,
    indicating via GDK_SUCCEED or GDK_FAIL whether the function succeeded or
    failed, M5/MAL now expects the function the return a str, either
    indication success via MAL_SUCCEED, or raising and error via
    "throwMessage(str, str);".
    Furthermore, M4/MEL handed over BAT arguments (and return values) as
    BAT* respectively BAT**. M5/MAL now simply ships the BAT's ID as int* in
    both directions. The function itself has to get the respective BAT
    descriptor, and take care of the proper reference management (was done
    by MEL in M4).
    Finally, the handling of variable argument lists is no longer done via the
    varargs/stdarg standard, but by direct access to the MAL interpreter
    stack; see the code for details.

    Following the "minimal-invasive" approach as proposed above, we prefer
    not the modify the M4 function signatures, but rather to add shallow
    wrapper around them that map the M5 signatures to the M4 signatures.
    Preferably, these wrappers are (i) gathered in once consecutive block in
    the code and (ii) the M5 wrapper for the M4 function
    		"M4_<mod>_<func>"
    is called
    		"M5_<mod>_<func>".

3c) Add/extend header file with exported prototype declarations.
    Without MEL, also the exporting of the prototype of externally callable
    functions (basically all functions that implement MAL commands) need to
    be provided by hand. Next to defining the respective "<mod_name>_macro"
		#ifdef _MSC_VER
		#ifndef LIB<MOD_NAME>
		#define <mod_name>_export extern __declspec(dllimport)
		#else
		#define <mod_name>_export extern __declspec(dllexport)
		#endif
		#else
		#define <mod_name>_export extern
		#endif
    We also need to export all prototypes as follows:
    		<mod_name>_export M5_<mod>_<func>(...);

4d) Porting the respective MIL test scripts to MAL.


Well, that's about it.

It took me about two complete nights to port the radix module from M4 to M5
--- well, most of it; I gave up on the atom definitions and the respective
multi-plexed cast functions. However, they're not core functionality of the
radix module, but were "merely" used to compare a MonetDB-like DSM approach
to a "standard" NSM approach. I left the respective code unchanged, hence,
it can be ported on need.


The diffs that come with these checkin messages show you to details what I
did. Feel free to lean from it. I tried to keep things as simple and generic
as possible.

In case you're curious, whether the goal of "minimal differences" was
reached, feel free to check it via (e.g.)

diff -du MonetDB/src/modules/plain/radix.mx monet5/src/modules/mal/radix.mx | cdiff


I hope, someone of you will find this story somehow informative...