This checkin is related to bug #1365379:
Until now, the "stable order by" information has not been
propagated from the parse tree to the XQuery Core tree.
Obviously, this prevented to exploit the information for
efficient processing.

The information is now propagated into the Core tree.  Both
branches (milprint_summer and algebra), however, still don't
respect it, but always implement the more strict "stable"
ordering.  While this is correct with respect to XQuery
semantics, it may be wasted work if "stable" has not been
requested by the user.

Exploiting the "stable" flag in the algebra branch could
result in a relaxed order constraint in the back-mapping step
of XQuery FLWOR expressions.  Our current understanding of the
"row numbering" operator, however, depends on a well-defined
order specification, which the relaxed ordering would violate.