Hi Stefan,
A belated big thanks for your efforts to see whether or not the kind-indices make sense!

Your results (fully below) showed indeed degradation if instead of using the kind-indices as candidate-list, a filter is applied (either beforehand by creating a candidate list on-the-fly) or afterwards (by filtering the result of ll-child/ll-descendant).

However, what I proposed was to do the kind-test inside the staircase join (not before or after), but inside it. That is, just before producing a result, look up kind[pre] and check if it is as desired.

I implemented this and checked it in. We thus have a new third variant of the ll-child/descendant(-or-self) algorithms:
(1) without any index
(2) with a sorted candidate-bat that contains PREs (=> extra skipping using this list!)
(3) NEW! with a non-nil kind_test parameter. As candidate-bat we pass PRE_KIND.

The results (110MB XMark, non-optimized gcc 64-bits oid32 compile, times in ms on db2.ins.cwi.nl) follow. As my new code introduces an extra if-statement in the (1) variant (no index), this variant might become slower. To test the impact of this, I benchmarked the default MXQ 'Before' and 'After' my modification. Note that the 'After' column contains the same default behaviour (using kind-indices as candidate-list, not a kind-test).

The third column is when all staircase-joins with kind-filters are handled with the new algorithm (3):

    Before After KindTest
q01   476   499   471
q02   805   839   795
q03  3039  3124  2976
q04  1086  1124  1083
q05   334   344   323
q06   142   149   139
q07   195   205   193
q08  1428  1453  1404
q09  1948  1941  1872
q10 12159 12175 11852
q11 10611 10647 10503
q12  5900  5927  5860
q13   608   629   598
q14  1904  2014  1876
q15   726   742   713
q16   614   627   602
q17   484   495   466
q18   282   290   268
q19  1138  1167  1093
q20  1491  1530  1476

We see a small degradation of overall performance caused by algorithm (1) as a result of my change, but the effect seems to be limited (3% perf degradation).

On the other hand if we refrain from using kind-indices as candidate-lists for scj, but use the kind-test instead, query performance does not degrade anymore like in your initial results (which shoed a factor 2-7 degradation). One can even claim a 1-2% performance improvement!

The problem with the kind-indices is that they are not selective. Text and element nodes are usually split (1/3 text, 2/3 elements) and such selectivities makes using an index not too useful, such that an integrated kind-test can beat it.

Given the effort that would be needed to keep these kind-indices up-to-date under updates, the decision to keep using them becomes a no-brainer. I just *removed* them from my update code-base (that is not checked in yet).

The new version (3) of child/descendant *is* checked in and you can play with it by setting the global variable (kind_idx := false;) in the MIL console (default is kind_idx := true -- no change of policy yet).

There is also another MIL variable (kind_debug) that when set to true prints some statistics on the costs of the staircase join operators. I include a trace produced by it on all XMark queries on the 110MB dataset below.

I want to draw your attention to the lines starting with nsloc_pre____ and the 'cands' statistic there, which is the size of a candidate list from a nsloc test (i.e. URI and local name, think //bidder, //person, etc.  in XMark.).

Our proposal for the update version is to keep a partial QN_NID BAT[oid,oid] index where we will only index those nsloc Qnames that occur M times in the document iff Mlog(M) < N (N = number of nodes).

The index must store NIDs as we cannot possible maintain PREs (because of massive shifts by each update) nor even RIDs (still in the thousands of shifts per single update). Thus at run-time, NIDs need to be translated to RIDs (with a fetchjoin) and then swizzled, and then *sorted*. This materializes the entire candidate list twice (O(M)), and sorts it O(mlog(M)). Clearly this makes sense only for small lists (i.e. if Mlog(M) > N, then a scj without candidate-list will be better).

In case of 110 MB XMark N=46M, which leads to a M=360K. The good news is that for all Xmark queries, M<360K seems to be the case for all nsloc scj-filters (max seen is M=59K, and often much less).

Note that the QN_NID index will be created on-the-fly, when a document is loaded, as this BAT will be expensive to flush to disk. Like the standoff indices, it will not belong to the persistent schema anymore, but be cached from the doc_index (or rather collection_index) table, such that they can be re-used by subsequent and concurrent queries. The QN_NID index can even be sorted on [QN,NID] for read-only documents. This means that selections on QN will be a slice-select rather than a hash-select, thus a view, and thus be free, such that our XMark results will not degrade at all.


Peter