Added support for functions returning a table
ex:
	create function f (i int) returns TABLE (s str, i int)
		return TABLE ( select 'testing', i );
or
	create function f (i int) returns TABLE (s str, i int)
		EXTERNAL NAME fmonet;

In the later case 'fmonet' is a mil/mal function which returns a bat[str,bat].
The strings are the column names and the bats the columns.

For table-functions we needed 'complex' (compound) types, these are
administered in the 'tables/columns' tables. The tables table column
isview, is now changed into 'type'. And we have the types 0 table, 1 view
and 2 generated. (tables view is changed to only return tables and views,
no compound types).

Support for view as BATS is removed!

The env table is removed, its now implemented using a table function (env())

The m5 backend now uses a table function to obtain the users from the m5
level.