When doing a batcast, VALconvert is called for each value.  But if the source type is an extension type the value is freed by VALconvert (since it does in-place conversion, the old value has to be cleaned up to make space for the new).  This is bad because the value comes directly from the heap and so cannot be freed.
Fix this problem by introducing a version of VALconvert, called
VALconvert2 which does not do in-place conversion, so we can skip the
freeing up of the source.

The new test in decimal, introduced in the previous checkin, caused a
crash in certain circumstances (optimized compilation on FC4), so a
new test is not needed.