part 2/3 of finishing/fixing the ID/IDREF implementation:
To comply with XQuery semantics in multi-doc collections, the ID/IDREF identifiers
must be specific to each document in a collection. Our solution is to prefix
the ids by the document root NID (as a string, and followed by a '_').

While this was done in the shredder, the implementation of fn:id()
in milprint_summer was still lacking the root NID prefix.
I added it.
To save one case in the MIL code from OID to STR
i.e., "[str](OID)" instead of "[str]([wrd](OID))",
I also added a "@0" to the root NID prefix in the shredder.