fixed bugs in global sessions variable handeling.         (reimplemented, the global variable values are now inside the         sql-variable stack only) Added a test for these kinds of variables too (src/test/session_vars)
Added sql.zero_or_one(b) function which returns
        NULL in case b is empty
        b.fetch(0) in case b has one row
        and raises an exception when there are more results

We now pass a struct exp_kind through the sql_value_exp. This is needed
to know when we require a single row (single column) result.
The struct keeps exp type, exp cardinality and exp reduce. The type can
be value or predicate, the cardinality value, column, set or table.
When reduce is set the expression may reduce the current column size.

removed a lot of function definitions from sql.mx (there where allready
moved to the respective modules or not needed anymore because we now keep
the module with the function definition in sql)

sql functions are allways in the sql_cache module, this is now also kept
in the function definition. This change is used in the code generation of
the m5 backend (ie its simplified). This solves a related crash.

keep integer values in a lng in the parser now (although we didn't have
an overflow in any of the tests, it could well happen). Should solve
some compilation warnings too.