Parser support for the `execute at' syntax extension (XRPC).
Pathfinder's parser now recognizes expressions of the kind

  "execute" "at" "{" Expr "}" "{" FunctionCall "}"

where the expression in the first pair of braces specifies an
URI where the function call given in the second pair should
be evaluated at.  Note that, in order to avoid parser conflicts,
we did not directly use the syntax proposed in a recent paper
submission, but rather this one that adds the curly braces.

The system will try hard to convert the expression given in
the first pair of braces into a string.

An expression as shown above will lead to a Core tree situation
that uses the new node kind `xrpc':

                xrpc
               /    \
             URI   FunctionCall

It's now Jennie's task to support that in the back-end.

TODO: If *not* touched the code that is currently existing in
Pathfinder and uses the "old" `import rpc-module' syntax.
Jennie, will you rip out this old stuff, once you have implemented
back-end support for the new syntax?