Our configure setup currently tries very hard to set the compiler options correctly. Allthough it does a damn good job at it, the usual approach used by distributions is different, ie they like to set all the CFLAGS etc them selfs.
This checkin breaks (probably literly) this setup in peaces.
Using the --enable/disable-strict the old behaviour is still there
and this is the default (the default can be easily changed using the
dft_strict variable)

With in the enable strict block, I currently keep the X_CFLAGS Hacks including
the -c99 and alike (probably means the strict is still needed for pathfinder).

The XOPEN etc flags, which should always end up in the config file are moved
out.

The default bits is now not set but obtained via a compilation of sizeof(long).