in case newcoll is empty, [logger_add_bat](pf_logger, [bat](newcoll), newcoll);  in  shred_doc_base() fails with !ERROR: interpret: no matching MIL operator to '[logger_add_bat](logger, BAT[oid,BAT], BAT[oid,str])'. !MAYBE YOU MEAN: !logger_add_bat(logger, BAT[void,any::1], str) : int
--- looks like some BUG in the multiplex code to me ---

for the sake of a greener testweb, this fix/hack(?) avoids the call
[logger_add_bat](pf_logger, [bat](newcoll), newcoll);
in case newcoll is empty.
(brings us down from 259 to "only" 92 failing Pathfinder tests)

neteffect: we also avoid setting/waiting for the pf_wal lock in this case.


Sjoerd,
I hope you don't mind that I "messed-around" with you code!

Sjoerd, Niels, me (and/or someone else) should check this potential
BUG in the multiplex-code/-handling...!