fixing BUG #1562868 XQ: complex q fails with error in "batbat_lng_add_inplace":
trading in optimization (performance?) for correctness:

The optimization to omit the eveluation of empty if-then-else branches
is broken, and since noone has been able to fix it, yet, we simply
disable this optimization --- the performance "penalty" should not be
too big, since the inpu to the respective branch is empty.

See
http://sourceforge.net/tracker/index.php?func=detail&aid=1562868&group_id=56967&atid=482468
for details.