Mapi protocol version 6 and 7.
Because I don't want to implement proto 7 in M4, it now uses proto 6.
Beware that this version was reverred to by M5, which now uses proto 7.

Proto 7 is now implemented in the client layer, which means MapiClient
can make full advantage of the security features offered by M5, which
were used by JDBC for a long while already.
TODO: make a flag for MapiClient that allows to set the desired hash to
use in the challenge response, like JdbcClient has.

I check this in on the "current" branch, because the patch is too big to
go into stable without at least a discussion on this short term for the
release.  The patch is "essential" though, in the sense that the
protocol versions need to get bumped before the release!