[ 1601761 ] XQ: Dagstuhl shredder bug-fix not correct
this bug was fixed, but the fixed decreased performance. We need
to first create a small bat, and then reallocate it, in order to
get the bat memory mapped (this is done in re-allocation).

this fix goes in conjuction with some other MonetDB checkins
done yesterday, which improve usage of memory mapping on growing
bat (string) heaps.