Fixed bug 1595701: "replace value" concatenates value iso. replacing it. The problem was that the replace value was translated to a delete of all children of the node and an insert of the new value.  The children were calculated at compile time. When a second replace value on the same node was done, the delete didn't have any effect (the nodes were already deleted), but the insert did have an effect, resulting in two texts being inserted.
The new way of doing things is to translate the query into a single
replaceElementContent which figures out which nodes to delete
dynamically.

Jens made the compiler changes, I did the rest.

Also added a test which tests a combination of replace value and a
change (insert in this case) of a descendant which is deleted by the
replace.  (This test uncovered a bug inserialize which has already
been fixed.)