Disabled failing algebra tests in the XQuery_0-14 release branch, only.
Since the algebra version will not be supported officially in the MonetDB/XQuery 0.14 release,
the respective problems will only be fixed in the development trunk.
Hence, no need to test them in the XQuery_0-14 release branch.

For Sjoerd:
I'll update the repective CVS tags by hand
to ensure that these changes will not be propagated to the development trunk.