added test for BUG #1618802 "XQ: document handling bug?": (http://sourceforge.net/tracker/index.php?func=detail&aid=1618802&group_id=56967&atid=482468)
A simple query loads 100 tiny document via fn:doc(<file_uri>).
Running the query a second time (both without and without
stopping the Mserver in between) fn:doc() is supposed to
re-use the cached document as opposed to re-shredding them.
This seems to work fine, in case the number of documents is small.
However, when the number of documents exceed a certain limit,
in my case (64-bit FC4 Linux, 64-bit OIDs, xquery_cacheMB=100)
with more than 95 of these tiny documents
(total serialized size is 404 KB, total size in dbfarm is ~500 KB),
MonetDB/XQuery seems to assume that the default 100 MB(!) cache
is full, and hence apparently tries to clear it before proceeding
(although actually access only cached documents, i.e., no new
documents have to be added to the cache).
Clearing the cache then fails (as reported in BUG #1618802) with
!ERROR: _del_doc(/net/corona.ins.cwi.nl/export/scratch0/manegold/Monet/Testing/Stable/.GNU.64.64.d.--disable-debug_--enable-assert_--enable-optimize.PREFIX./PATHFINDER/mTests/tests/BugTracker/document_handling.SF-1618802/95.xml): document not found in database (1 such errors)!